Move evolution and grammar modules under lib/ and update imports
