+++ list.py
@@ -7,7 +7,7 @@ import datetime
 
 from lib.ec import explorationCompression, commandlineArguments
 from lib.utilities import eprint, numberOfCPUs, flatten, testTrainSplit
-from grammar import Grammar
+from lib.grammar import Grammar
 from lib.task import Task
 from lib.type import Context, arrow, tbool, tlist, tint, t0, UnificationFailure
 from lib.domains.list.listPrimitives import basePrimitives, primitives, McCarthyPrimitives, bootstrapTarget_extra, no_length
