+++ puddleworld.py
@@ -8,7 +8,7 @@ import os
 import random
 
 from lib.ec import commandlineArguments
-from grammar import Grammar
+from lib.grammar import Grammar
 from lib.utilities import eprint, numberOfCPUs
 
 from lib.domains.puddleworld.puddleworldPrimitives import primitives
